pyproject.toml
[project]
name = "query-fanout-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.50.1",
    "numpy>=2.3.4",
    "plotly>=6.4.0",
    "kaleido>=0.2.1",
]

